+++ Kiran/Addition.java
@@ -1,0 +1,12 @@
+package Kiran;
+
+public class Addition {
+
+
+    public static void main(String[] args) {
+        int a = 3;
+        int b = 5;
+               int sum = a+b;
+               System.out.println("sum = " +sum);
+    }
+}
