--- KiranSutar/addition.java
@@ -1,13 +1,0 @@
-package KiranSutar;
-
-public class addition {
-
-    public static void main(String args[]) {
-
-
-        int a = 1;
-        int b = 1;
-        int sum = a + b;
-        System.out.println("addition = " +sum);
-    }
-}
